+++ src/d_deh.h
@@ -70,6 +70,7 @@ extern char     *s_PRESSYN;
 extern char     *s_PRESSA;
 extern char     *s_QUITMSG;
 extern char     *s_QLPROMPT;
+extern char     *s_DELPROMPT;
 extern char     *s_NIGHTMARE;
 extern char     *s_SWSTRING;
 extern char     *s_MSGOFF;