+++ docs/releasenotes.md
@@ -4,6 +4,7 @@
 * Frames from *DOOM’s* rocket launcher are no longer shown when firing the missile launcher in [*Freedoom*](http://freedoom.github.io/).
 * Entering the `idmypos` cheat will no longer cause a crash.
 * Homing rockets fired by revenants will no longer randomly become regular rockets, and vice-versa, if a game is saved while they are in mid-air. In order for this bug to be fixed, the format of savegames has changed. Savegames created using previous versions of *DOOM Retro* are not compatible with this version.
+* Pressing the <kbd>DEL</kbd> when in the savegame menu will now delete the currently selected savegame.
 
 ---